Allow an ACL option for objects to be uploaded
By default, I believe, when uploading a file using
`lambduh-put-s3-object` it will put the file up with the 'private' as
the ACL, with no ability to define an alternative policyw. This should
allow library users to define the ACL that files should be uploaded with
instead.

NOTE: Tests needed